fixed injection_gen issues, working on lc_gen
had to do a few workarounds to get the injections ingested correctly by nmma, seem to have an issue with simulation_id not being present as a key without manually adding it. Will have to look into that more.

bigger issue: ztf sampling and related args don't seem to work, and in addition to that, the Bu2019lm lightcurve generation is failing wiht "ValueError: X has 4 features, but GaussianProcessRegressor is expecting 2 features as input." seems to be an issue of sklearn versioning that I'm not sure to fix (lc_gen needs something like >1.2, but create_inj needs <1.0.2).

Possibly as a result of the ztf flag issues, the lightcurves that are generated are all inf magnitudes (regardless of which filter I use)