output lightcurves (but odd)
lightcurves were able to be output when using my laptop, but they seem to ignore my arguments provided in light_curve_generation for tmin and tmax. It also outputs in all filters rather than just the g filter. FInally, it outputs as a json file, which I've not encountered with light curve files in nmma before to my knowledge. But progress!