created bulk_data_generation script
this will allow for rapid deployment of injection generation, fitting, and chi2 calculation for a large dataset once the injection generation issue is solved.

In the process of writing this script, some of the other scripts were modified to accept additional arguments. Most significantly, there is now a check in analysis_script.py to continually check to see if all jobs are completed (will be complete if *result.json is found) or if the execution time exceeds some timeout value. Script will continue to run until all analyses are completed.

These changes should have been made in such a way that other uses of the scripts are unaffected.