prior/augmented injection tweaks
Changed the luminosity distance ranges for some of the priors to prevent overly close or overly far objects in the sample, also corrected the augmentation of points so there's actually a 0.5 day cadence and filters out the extra filters we don't want. Added a function to make a quick plot of injected lightcurves for validation.  corrected time range in lightcurve_analysis to account for new tmin and tmax applications.

Still seeing some issues with showing rise time, specifically for the GRB model (TrPi2018).