fixed bug with trigger time
empty ev.dat file was being caused by the trigger time being set to 0.1 rather than a time that would make sense to nmma (this is because the old json format used relative times, whereas the current trigger needs to be absolute). Added a function to grab the trigger time from the lightcurve file.