follow up functions.
functions to generate the first lightcurve file as well as a function to generate a follow-up lightcurve. the correct_times argument on both currently is not functioning in all edge-cases. Right now, it will read in the time of the first detection, so if the lowest time measurement is a non-detection, the times will be thrown off.