reworked model choices
argument in analysis_script for choosing models to fit no longer has limits by choice; instead, the priors directory is cross-referenced with the list of models provided, and any that do not have an associated prior are removed from the list of models to fit. Also added the Bu2022Ye model prior.