account for failed fits and add more stats
this will account for cases where a fit fails, wherein the likelihood and bf will be set to negative inf. this also adds the other best_fit json values to the lightcurve object so we can still access those after the best fit is deleted. seems that the function for moving the bestfit_json file is still not fully functional at the moment.